+++ manual/changelog.md
@@ -1,5 +1,15 @@
 English page is here [>> English page](changelog-en)
 
+version 4.9.0
+
+    Fixed method result bug.
+
+    Fixed Generics bug.
+
+    Fixed block object bug
+
+    Added ? and : operator for null checking and exception checking.
+
 version 4.8.9
 
     Fixed JIT bug